semantic_twin_engine/requirements.txt
# Semantic Twin Engine Dependencies
# ==================================

# Core
openai>=1.0.0
pydantic>=2.0.0
pyyaml>=6.0
python-dotenv>=1.0.0

# Math & Vectors
numpy>=1.24.0
scikit-learn>=1.3.0

# Resilience
tenacity>=8.2.0
